requirements.txt
llama-index
openai
pypdf
python-dotenv
streamlit